requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-community>=0.2.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
pymupdf>=1.24.0
ollama>=0.2.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.18.0
